Finalized Autonomous procedures. Improved ease of use with an incomplete field. Combined Vuforia utilities/declarations/definitions into a single file. Fixed Point utility class

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Utilities/Size.java
@@ -1,10 +1,10 @@
 package org.firstinspires.ftc.teamcode.Utilities;
 
 public class Size {
-    public double width;
-    public double height;
+    public float width;
+    public float height;
 
-    public Size(double width, double height) {
+    public Size(float width, float height) {
         this.width = width;
         this.height = height;
     }
